Add human readable labels to a table

cbs4_get_observations() downloads a data table in which the Measure and dimension columns contain Codes from the Identifier columns of MeasureCodes and <Dimension>Codes data.frames that can be found in the metadata downloaded with cbs4_get_metadata().

cbs4_add_label_columns(data, ...)

Arguments

data
downloaded with cbs4_get_data()

...
not used

Details

cbs4_get_observations() adds for the Measure and each <Dimension> column an extra column MeasureLabel ( <Dimension>Label) that contains the Title of each code.

See also

Examples

# works on observations
obs <- cbs4_get_observations("84287NED", Perioden="2019MM12")
obs # without label columns

#>     Id Measure ValueAttribute Value Perioden BedrijfstakkenBranchesSBI2008
#> 1 1064 M002042           None  0.23 2019MM12                        300007
#> 2  797 M002042           None  0.22 2019MM12                        350000
#> 3  530 M002042           None  0.07 2019MM12                        307500
#> 4  263 M002042           None  0.21 2019MM12                       T001081

obs_labeled <- cbs4_add_label_columns(obs)
obs_labeled

#>     Id Measure       MeasureLabel ValueAttribute Value Perioden PeriodenLabel
#> 1 1064 M002042 Vacature-indicator           None  0.23 2019MM12 2019 december
#> 2  797 M002042 Vacature-indicator           None  0.22 2019MM12 2019 december
#> 3  530 M002042 Vacature-indicator           None  0.07 2019MM12 2019 december
#> 4  263 M002042 Vacature-indicator           None  0.21 2019MM12 2019 december
#>   BedrijfstakkenBranchesSBI2008 BedrijfstakkenBranchesSBI2008Label
#> 1                        300007    G-N Commerciële dienstverlening
#> 2                        350000                   F Bouwnijverheid
#> 3                        307500                        C Industrie
#> 4                       T001081  A-U Alle economische activiteiten

# works on data
d <- cbs4_get_data("84287NED", Perioden="2019MM12")
d # cbs4_get_data automagically labels measure columns.

#>   Perioden BedrijfstakkenBranchesSBI2008 Vacature-indicator
#> 1 2019MM12                        300007               0.23
#> 2 2019MM12                        307500               0.07
#> 3 2019MM12                        350000               0.22
#> 4 2019MM12                       T001081               0.21

d_labeled <- cbs4_add_label_columns(d)
d_labeled

#>   Perioden PeriodenLabel BedrijfstakkenBranchesSBI2008
#> 1 2019MM12 2019 december                        300007
#> 2 2019MM12 2019 december                        307500
#> 3 2019MM12 2019 december                        350000
#> 4 2019MM12 2019 december                       T001081
#>   BedrijfstakkenBranchesSBI2008Label Vacature-indicator
#> 1    G-N Commerciële dienstverlening               0.23
#> 2                        C Industrie               0.07
#> 3                   F Bouwnijverheid               0.22
#> 4  A-U Alle economische activiteiten               0.21